reorganizing and using EuclideanCollections to instantiate point maps
